+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Added support for Latvian 🇱🇻! This includes the Latvian part of the reading comprehension
+  dataset MultiWikiQA, the sentiment classification dataset Latvian Twitter Sentiment,
+  the linguistic acceptability dataset ScaLA-lv, the NER dataset WikiANN-lv, the
+  NER dataset FullStack-NER-lv, and the knowledge dataset MMLU-lv.
 - It is now possible to evaluate how much a model adhere to European values! 🇪🇺 This
   probes 53 questions from the European values survey, which have been chosen based on
   an optimisation procedure that maximises agreement across the EU. We then measure how